% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphereML.R
\name{sphereML}
\alias{sphereML}
\title{A shiny application of sphereML}
\usage{
sphereML()
}
\value{
A user interface of shiny application.
}
\description{
In this package, machine learning based analysis can be performed to predict students' performance outcomes in physics.
}
\examples{
library(sphereML)
sphereML()
}
