% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_sphereML.R
\name{start_sphereML}
\alias{start_sphereML}
\title{Start a shiny application of sphereML}
\usage{
start_sphereML()
}
\value{
A user interface of shiny application.
}
\description{
An interactive Shiny application for running a machine learning analysis to the 'spheredata' package.
}
\details{
This starts the application on the users local computer.
}
\examples{
\dontrun{
library(sphereML)
start_sphereML()
}
}
